ultimatevocalremovergui/requirements.txt
altgraph==0.17.3
audioread==3.0.0
certifi==2022.12.07
cffi==1.15.1
cryptography==3.4.6
einops==0.6.0
future==0.18.3
julius==0.2.7
kthread==0.2.3
matchering==2.0.6
ml_collections==0.1.1
natsort==8.2.0
omegaconf==2.2.3
opencv-python==4.6.0.66
Pillow==9.3.0
psutil==5.9.4
pydub==0.25.1
pyglet==1.5.23
pyperclip==1.8.2
pyrubberband==0.3.0
pytorch_lightning==2.0.0
PyYAML==6.0
resampy==0.4.2
soundstretch==1.2
wget==3.2
samplerate==0.1.0
screeninfo==0.8.1
diffq==0.2.3
onnx==1.13.0
onnxruntime
onnx2pytorch==0.4.1
SoundFile==0.11.0; sys_platform != 'darwin'
PySoundFile==0.9.0.post1; sys_platform == 'darwin'
